requirements.txt
asgiref==3.7.2
Django==5.0
django-appconf==1.0.6
django-select2==8.1.2
django-tables2==2.7.0
Pillow==10.1.0
sqlparse==0.4.4
tzdata==2023.3
python-docx==1.1.0

dramatiq~=1.15.0
django-filter~=23.5